Money Converter Program
Assignment3 at my study. This study case is decided to convert a kind of currency. For example IDR to USD and JPY, USD to IDR and JPY, and JPY to IDR and USD. So I will make a program to convert kind of currency. The user needs to choose one kind of currency, then the user input how much money, after that, it will be converted to another kind of currency.