Family Catering Program
Family Catering in Jogja, located in the Tamsis area, requires the assistance of a Cashier Program to facilitate the data collection of Sales Transactions, which is expected to help reduce human error related to price calculation errors. Every Buyer who orders catering will be recorded the transaction number in the general format specified is _ _ _ _ _ DDMMYY, where the first 5 digits are the transaction number (Example: 00001301120). Next, the name of the buyer will also be completed. Each buyer can order various types of food in one transaction, according to their needs 
and desires. For each type of food or menu, the data that will be inputted such as Menu Code,  Menu Name, Price Per Serving, How Many Portions to be purchased, and then the total is calculated for each Menu Item automatically. The menu listing process will be repeated as many as the number of menus desired by the Buyer. The last process required, as with any cashier program in general, is to calculate the total payment, discount, and change. Where Family Catering provides a discount of 2% for every purchase above Rp. 500,000.00 (applies multiples). The program will display Total paid, Discount Repetition, Discount Amount, and Total Paid after discount. Then, the Input Paid 
Money will proceed along with checking whether the money paid is minus or not, and the change will be automatically calculated. The menu list notes can be seen in the table below. And to be able to enter the Cashier program begins with User Verification with Username and Password = "admin".

I upgrade the program, when input the code of menu from the table, it will automatically input the name and price. But, when we input the code not from the table, it will not 
automatically input the data. I also add menu check.